# src/main/resources/openapi-GENERATED.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/devices/{deviceId}/name:
    put:
      tags:
      - device-controller
      operationId: updateDeviceName
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeviceEntity"
  /api/telemetry:
    post:
      tags:
      - telemetry-controller
      operationId: ingestTelemetry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryIngestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TelemetryResponse"
    get:
      tags:
      - telemetry-controller
      operationId: fetchTelemetry
      parameters:
      - name: deviceId
        in: query
        required: true
        schema:
          type: string
      - name: since
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: end
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: tripId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelemetryResponse"
  /api/telemetry/vehicle-stats:
    get:
      tags:
      - telemetry-controller
      operationId: getVehicleStats
      parameters:
      - name: deviceId
        in: query
        required: true
        schema:
          type: string
      - name: since
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: timeBetweenDrivesInSeconds
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 600
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/telemetry/trips-per-weekday:
    get:
      tags:
      - telemetry-controller
      operationId: getTripsPerWeekday
      parameters:
      - name: deviceId
        in: query
        required: true
        schema:
          type: string
      - name: since
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: timeBetweenDrivesInSeconds
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 600
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int32
  /api/telemetry/latest:
    get:
      tags:
      - telemetry-controller
      operationId: fetchLatestTelemetry
      parameters:
      - name: deviceId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TelemetryResponse"
  /api/telemetry/trips:
    get:
      tags:
      - telemetry-controller
      operationId: fetchTelemetryGroupedByTrip
      parameters:
      - name: deviceId
        in: query
        required: true
        schema:
          type: string
      - name: since
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: end
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: timeBetweenTripsInSeconds
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1800
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: "#/components/schemas/TelemetryResponse"
  /api/devices:
    get:
      tags:
      - device-controller
      operationId: getAllDevices
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceEntity"
  /api/trips/{tripId}:
    patch:
      tags:
      - trip-controller
      operationId: updateTrip
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TripResponse"
components:
  schemas:
    DeviceEntity:
      type: object
      properties:
        deviceId:
          type: string
        name:
          type: string
    TelemetryIngestRequest:
      type: object
      properties:
        deviceId:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        aggregated_data:
          type: object
          additionalProperties:
            type: object
        timed_data:
          type: object
          additionalProperties:
            type: object
        errors:
          type: object
          additionalProperties:
            type: object
      required:
      - deviceId
      - start_time
    TelemetryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        tripId:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        timed_data:
          type: object
          additionalProperties:
            type: object
        aggregated_data:
          type: object
          additionalProperties:
            type: object
    TripUpdateRequest:
      type: object
      properties:
        startLocation:
          type: string
        endLocation:
          type: string
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        startLocation:
          type: string
        endLocation:
          type: string